Removed useless files and organized the codebase
