+++ src/lib/index.js
@@ -1,1 +1,31 @@
 // place files you want to import through the `$lib` alias in this folder.
+
+// Export common utilities and types for easier imports
+
+// Types
+export * from './types';
+
+// Store exports
+export { todos } from './stores/todos';
+export { 
+  notifications, 
+  addNotification, 
+  markAsRead, 
+  markAllAsRead, 
+  unreadCount,
+  connected,
+  initializeSocket,
+  joinGroup,
+  cleanup,
+  sendDirectGroupInvitation
+} from './stores/socket';
+export { 
+  toasts,
+  addToast as showToast, 
+  removeToast as dismissToast 
+} from './stores/toasts';
+export { 
+  addOnlineUser,
+  removeOnlineUser,
+  updateOnlineUsers
+} from './stores/onlineUsers';
