--- src/lib/types/user.ts
@@ -1,8 +1,0 @@
-export interface User {
-	_id: string;
-	username: string;
-	email: string;
-	profileImage?: string;
-	createdAt?: string;
-	updatedAt?: string;
-}
